(BSR) feat(user): move useAvaibledCredit in shared/user
