+++ src/features/bookOffer/pages/BookingConfirmation.native.test.tsx
@@ -13,7 +13,7 @@ import { BookingConfirmation } from './BookingConfirmation'
 jest.mock('react-query')
 jest.mock('features/offer/api/useOffer')
 
-jest.mock('features/user/helpers/useAvailableCredit', () => ({
+jest.mock('shared/user/useAvailableCredit', () => ({
   useAvailableCredit: jest.fn(() => ({ isExpired: false, amount: 2000 })),
 }))
