+++ src/features/bookOffer/pages/BookingConfirmation.tsx
@@ -5,12 +5,12 @@ import styled from 'styled-components/native'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useShareOffer } from 'features/offer/helpers/useShareOffer'
-import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { analytics } from 'libs/firebase/analytics'
 import { useShowReview } from 'libs/hooks/useShowReview'
 import { formatToFrenchDecimal } from 'libs/parsers'
 import { BatchEvent, BatchUser } from 'libs/react-native-batch'
 import { WebShareModal } from 'libs/share/WebShareModal'
+import { useAvailableCredit } from 'shared/user/useAvailableCredit'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
 import { ButtonTertiaryPrimary } from 'ui/components/buttons/ButtonTertiaryPrimary'
