+++ src/features/favorites/components/Favorite.native.test.tsx
@@ -7,10 +7,10 @@ import { api } from 'api/api'
 import { ExpenseDomain, FavoriteResponse, UserProfileResponse } from 'api/gen'
 import { initialFavoritesState } from 'features/favorites/context/reducer'
 import { favoriteResponseSnap as favorite } from 'features/favorites/fixtures/favoriteResponseSnap'
-import { Credit } from 'features/user/helpers/useAvailableCredit'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
 import { analytics } from 'libs/firebase/analytics'
+import { Credit } from 'shared/user/useAvailableCredit'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
 import { act, fireEvent, render, cleanup, waitFor } from 'tests/utils'
