+++ src/features/home/components/headers/HomeHeader.tsx
@@ -9,11 +9,11 @@ import { useBeneficiaryValidationNavigation } from 'features/auth/helpers/useBen
 import { GeolocationBanner } from 'features/home/components/GeolocationBanner'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { isUserBeneficiary } from 'features/profile/helpers/isUserBeneficiary'
-import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
-import { useGetDepositAmountsByAge } from 'features/user/helpers/useGetDepositAmountsByAge/useGetDepositAmountsByAge'
 import { env } from 'libs/environment'
 import { useGeolocation, GeolocPermissionState } from 'libs/geolocation'
 import { formatToFrenchDecimal } from 'libs/parsers'
+import { useAvailableCredit } from 'shared/user/useAvailableCredit'
+import { useGetDepositAmountsByAge } from 'shared/user/useGetDepositAmountsByAge'
 import { theme } from 'theme'
 import { PageHeader } from 'ui/components/headers/PageHeader'
 import { BannerWithBackground } from 'ui/components/ModuleBanner/BannerWithBackground'
