+++ src/features/identityCheck/pages/Stepper.tsx
@@ -14,9 +14,9 @@ import { useSetSubscriptionStepAndMethod } from 'features/identityCheck/useSetCu
 import { useSubscriptionSteps } from 'features/identityCheck/useSubscriptionSteps'
 import { getStepState } from 'features/identityCheck/utils/getStepState'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { hasOngoingCredit } from 'features/user/helpers/useAvailableCredit'
 import { amplitude } from 'libs/amplitude'
 import { analytics } from 'libs/firebase/analytics'
+import { hasOngoingCredit } from 'shared/user/useAvailableCredit'
 import { ButtonTertiaryWhite } from 'ui/components/buttons/ButtonTertiaryWhite'
 import { Li } from 'ui/components/Li'
 import { useModal } from 'ui/components/modals/useModal'
