+++ src/features/identityCheck/pages/confirmation/IdentityCheckHonor.tsx
@@ -11,9 +11,9 @@ import { Declaration } from 'features/identityCheck/atoms/Declaration'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { useSubscriptionNavigation } from 'features/identityCheck/useSubscriptionNavigation'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { hasOngoingCredit } from 'features/user/helpers/useAvailableCredit'
 import { amplitude } from 'libs/amplitude'
 import { QueryKeys } from 'libs/queryKeys'
+import { hasOngoingCredit } from 'shared/user/useAvailableCredit'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { useEnterKeyAction } from 'ui/hooks/useEnterKeyAction'
