+++ src/features/search/pages/modals/PriceModal/PriceModal.tsx
@@ -16,9 +16,9 @@ import { useSearch } from 'features/search/context/SearchWrapper'
 import { MAX_PRICE } from 'features/search/helpers/reducer.helpers'
 import { makeSearchPriceSchema } from 'features/search/helpers/schema/makeSearchPriceSchema/makeSearchPriceSchema'
 import { SearchState, SearchView } from 'features/search/types'
-import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { analytics } from 'libs/firebase/analytics'
 import { formatToFrenchDecimal } from 'libs/parsers'
+import { useAvailableCredit } from 'shared/user/useAvailableCredit'
 import { Banner } from 'ui/components/Banner'
 import { Form } from 'ui/components/Form'
 import { InputError } from 'ui/components/inputs/InputError'
